Documentation improvement and minor refactoring.
