+++ src/main/java/com/github/dakusui/pcond/core/Evaluable.java
@@ -1,6 +1,6 @@
 package com.github.dakusui.pcond.core;
 
-import com.github.dakusui.pcond.core.context.CurriedContext;
+import com.github.dakusui.pcond.experimentals.currying.context.CurriedContext;
 
 import java.util.List;
 import java.util.Optional;
