+++ src/main/java/com/github/dakusui/pcond/experimentals/currying/context/package-info.adoc
