+++ src/main/java/com/github/dakusui/pcond/experimentals/currying/multi/package-info.adoc
