+++ src/main/java/com/github/dakusui/pcond/experimentals/currying/package-info.adoc
