+++ src/main/java/com/github/dakusui/pcond/fluent/with-dummyvalue-filter.txt
@@ -1,0 +1,59 @@
+EXPECTED:
+    "hello"                                   ->WHEN:transform          ->true
+                                              ->    length              ->5
+    5                                         ->  THEN:>[1]             ->true
+    Book:[title:<De Bello G...i appellantur.>]->WHEN:allOf              ->(not evaluated)
+                                              ->  transform:title       ->"De Bello Gallico"
+    "De Bello Gallico"                        ->  THEN:allOf            ->(not evaluated)
+                                              ->      isNotNull         ->true
+[0]                                           ->    transform:parseInt  ->(not evaluated)
+    NumberFormatException:"For input s...ico""->    THEN:allOf          ->(not evaluated)
+                                              ->        >=[10]          ->(not evaluated)
+                                              ->        <[40]           ->(not evaluated)
+    Book:[title:<De Bello G...i appellantur.>]->  transform:abstractText->"Gallia est omnis divis...li appellantur."
+    "Gallia est omnis divis...li appellantur."->  THEN:allOf            ->true
+                                              ->      not:isNull        ->true
+                                              ->    transform:length    ->145
+    145                                       ->    THEN:allOf          ->true
+[1]                                           ->        >=[200]         ->true
+                                              ->        <[400]          ->true
+
+.Detail of failure [0]
+----
+parseInt
+----
+
+.Detail of failure [1]
+----
+>=[200]
+----
+
+ACTUAL:
+    "hello"                                   ->WHEN:transform          ->true
+                                              ->    length              ->5
+    5                                         ->  THEN:>[1]             ->true
+    Book:[title:<De Bello G...i appellantur.>]->WHEN:allOf              ->(not evaluated)
+                                              ->  transform:title       ->"De Bello Gallico"
+    "De Bello Gallico"                        ->  THEN:allOf            ->(not evaluated)
+                                              ->      isNotNull         ->true
+[0]                                           ->    transform:parseInt  ->(not evaluated)
+    NumberFormatException:"For input s...ico""->    THEN:allOf          ->(not evaluated)
+                                              ->        >=[10]          ->(not evaluated)
+                                              ->        <[40]           ->(not evaluated)
+    Book:[title:<De Bello G...i appellantur.>]->  transform:abstractText->"Gallia est omnis divis...li appellantur."
+    "Gallia est omnis divis...li appellantur."->  THEN:allOf            ->true
+                                              ->      not:isNull        ->true
+                                              ->    transform:length    ->145
+    145                                       ->    THEN:allOf          ->true
+[1]                                           ->        >=[200]         ->true
+                                              ->        <[400]          ->true
+
+.Detail of failure [0]
+----
+parseInt
+----
+
+.Detail of failure [1]
+----
+>=[200]
+----
