+++ src/main/java/com/github/dakusui/pcond/fluent/without-dummyvalue-filter.txt
@@ -1,0 +1,91 @@
+EXPECTED:
+    ["hello",Book:[title:<De Bello G...i appellantur.>]]->allOf                     ->(not evaluated)
+                                                        ->  transform:at[0]         ->"hello"
+    "hello"                                             ->  WHEN:transform          ->true
+                                                        ->      length              ->5
+    5                                                   ->    THEN:>[1]             ->true
+    ["hello",Book:[title:<De Bello G...i appellantur.>]]->  transform:at[1]         ->Book:[title:<De Bello G...i appellantur.>]
+    Book:[title:<De Bello G...i appellantur.>]          ->  WHEN:allOf              ->(not evaluated)
+                                                        ->    transform:title       ->"De Bello Gallico"
+    "De Bello Gallico"                                  ->    THEN:allOf            ->(not evaluated)
+                                                        ->        isNotNull         ->true
+[0]                                                     ->      transform:parseInt  ->(not evaluated)
+    NumberFormatException:"For input s...ico""          ->      THEN:allOf          ->(not evaluated)
+                                                        ->          >=[10]          ->(not evaluated)
+                                                        ->          <[40]           ->(not evaluated)
+    Book:[title:<De Bello G...i appellantur.>]          ->    transform:abstractText->"Gallia est omnis divis...li appellantur."
+    "Gallia est omnis divis...li appellantur."          ->    THEN:allOf            ->true
+                                                        ->        not:isNull        ->true
+                                                        ->      transform:length    ->145
+    145                                                 ->      THEN:allOf          ->true
+[1]                                                     ->          >=[200]         ->true
+                                                        ->          <[400]          ->true
+
+.Detail of failure [0]
+----
+parseInt
+----
+ACTUAL:
+    ["hello",Book:[title:<De Bello G...i appellantur.>]]->allOf                     ->(not evaluated)
+                                                        ->  transform:at[0]         ->"hello"
+    "hello"                                             ->  WHEN:transform          ->true
+                                                        ->      length              ->5
+    5                                                   ->    THEN:>[1]             ->true
+    ["hello",Book:[title:<De Bello G...i appellantur.>]]->  transform:at[1]         ->Book:[title:<De Bello G...i appellantur.>]
+    Book:[title:<De Bello G...i appellantur.>]          ->  WHEN:allOf              ->(not evaluated)
+                                                        ->    transform:title       ->"De Bello Gallico"
+    "De Bello Gallico"                                  ->    THEN:allOf            ->(not evaluated)
+                                                        ->        isNotNull         ->true
+[0]                                                     ->      transform:parseInt  ->NumberFormatException:"For input s...ico""
+    NumberFormatException:"For input s...ico""          ->      THEN:allOf          ->(not evaluated)
+                                                        ->          >=[10]          ->(not evaluated)
+                                                        ->          <[40]           ->(not evaluated)
+    Book:[title:<De Bello G...i appellantur.>]          ->    transform:abstractText->"Gallia est omnis divis...li appellantur."
+    "Gallia est omnis divis...li appellantur."          ->    THEN:allOf            ->false
+                                                        ->        not:isNull        ->true
+                                                        ->      transform:length    ->145
+    145                                                 ->      THEN:allOf          ->false
+[1]                                                     ->          >=[200]         ->false
+                                                        ->          <[400]          ->true
+
+.Detail of failure [0]
+----
+Input: 'De Bello Gallico'
+Input Type: java.lang.String
+Thrown Exception: 'java.lang.NumberFormatException'
+Exception Message: > For input string: "De Bello Gallico"
+	java.lang.NumberFormatException.forInputString(NumberFormatException.java:65)
+	java.lang.Integer.parseInt(Integer.java:580)
+	java.lang.Integer.parseInt(Integer.java:615)
+	...internal.package.InternalClass.internalMethod(InternalClass.java:0)
+	sun.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
+	sun.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:62)
+	sun.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
+	java.lang.reflect.Method.invoke(Method.java:498)
+	org.junit.runners.model.FrameworkMethod$1.runReflectiveCall(FrameworkMethod.java:59)
+	org.junit.internal.runners.model.ReflectiveCallable.run(ReflectiveCallable.java:12)
+	org.junit.runners.model.FrameworkMethod.invokeExplosively(FrameworkMethod.java:56)
+	org.junit.internal.runners.statements.InvokeMethod.evaluate(InvokeMethod.java:17)
+	org.junit.internal.runners.statements.RunBefores.evaluate(RunBefores.java:26)
+	org.junit.internal.runners.statements.RunAfters.evaluate(RunAfters.java:27)
+	org.junit.runners.ParentRunner$3.evaluate(ParentRunner.java:306)
+	org.junit.runners.BlockJUnit4ClassRunner$1.evaluate(BlockJUnit4ClassRunner.java:100)
+	org.junit.runners.ParentRunner.runLeaf(ParentRunner.java:366)
+	org.junit.runners.BlockJUnit4ClassRunner.runChild(BlockJUnit4ClassRunner.java:103)
+	org.junit.runners.BlockJUnit4ClassRunner.runChild(BlockJUnit4ClassRunner.java:63)
+	org.junit.runners.ParentRunner$4.run(ParentRunner.java:331)
+	org.junit.runners.ParentRunner$1.schedule(ParentRunner.java:79)
+	org.junit.runners.ParentRunner.runChildren(ParentRunner.java:329)
+	org.junit.runners.ParentRunner.access$100(ParentRunner.java:66)
+	org.junit.runners.ParentRunner$2.evaluate(ParentRunner.java:293)
+	org.junit.runners.ParentRunner$3.evaluate(ParentRunner.java:306)
+	org.junit.runners.ParentRunner.run(ParentRunner.java:413)
+	org.junit.runner.JUnitCore.run(JUnitCore.java:137)
+	...internal.package.InternalClass.internalMethod(InternalClass.java:0)
+
+----
+
+.Detail of failure [1]
+----
+145
+----
