+++ src/test/java/com/github/dakusui/pcond/propertybased/tests/CurriedContextPredicateTest.java
@@ -1,6 +1,6 @@
 package com.github.dakusui.pcond.propertybased.tests;
 
-import com.github.dakusui.pcond.forms.Experimentals;
+import com.github.dakusui.pcond.experimentals.currying.CurriedFunctions;
 import com.github.dakusui.pcond.propertybased.utils.PropertyBasedTestBase;
 import com.github.dakusui.pcond.propertybased.utils.TestCase;
 import com.github.dakusui.pcond.propertybased.utils.TestCaseParameter;
@@ -9,7 +9,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.junit.runners.Parameterized.Parameters;
 
-import static com.github.dakusui.pcond.forms.Experimentals.toCurriedContextStream;
+import static com.github.dakusui.pcond.experimentals.currying.CurriedFunctions.toCurriedContextStream;
 import static com.github.dakusui.pcond.forms.Functions.streamOf;
 import static com.github.dakusui.pcond.forms.Predicates.*;
 import static com.github.dakusui.pcond.propertybased.utils.TestCheck.equalsPredicate;
@@ -33,7 +33,7 @@ public class CurriedContextPredicateTest extends PropertyBasedTestBase {
         v = "hello",
         transform(streamOf()                                        // (1)
             .andThen(toCurriedContextStream()))                            // (2)
-            .check(anyMatch(Experimentals.toCurriedContextPredicate(isNotNull()))),
+            .check(anyMatch(CurriedFunctions.toCurriedContextPredicate(isNotNull()))),
         Object.class)
         .addExpectationPredicate(equalsPredicate(v))
         .build();
