+++ src/test/java/com/github/dakusui/pcond/propertybased/tests/StreamNoneMatchPredicateTest.java
@@ -1,6 +1,6 @@
 package com.github.dakusui.pcond.propertybased.tests;
 
-import com.github.dakusui.pcond.forms.Experimentals;
+import com.github.dakusui.pcond.experimentals.currying.CurriedFunctions;
 import com.github.dakusui.pcond.forms.Functions;
 import com.github.dakusui.pcond.forms.Printables;
 import com.github.dakusui.pcond.propertybased.utils.PropertyBasedTestBase;
@@ -16,8 +16,8 @@ import org.junit.runners.Parameterized.Parameters;
 import java.util.function.Function;
 import java.util.stream.Stream;
 
-import static com.github.dakusui.pcond.forms.Experimentals.nest;
-import static com.github.dakusui.pcond.forms.Experimentals.toCurriedContextPredicate;
+import static com.github.dakusui.pcond.experimentals.currying.CurriedFunctions.nest;
+import static com.github.dakusui.pcond.experimentals.currying.CurriedFunctions.toCurriedContextPredicate;
 import static com.github.dakusui.pcond.forms.Functions.length;
 import static com.github.dakusui.pcond.forms.Predicates.*;
 import static com.github.dakusui.pcond.propertybased.utils.TestCheck.*;
@@ -45,7 +45,7 @@ public class StreamNoneMatchPredicateTest extends PropertyBasedTestBase {
         Stream.of(null, "Hi", "hello", "world", null),
         transform(nest(asList("1", "2", "o")))
             .check(noneMatch(
-                Experimentals.toCurriedContextPredicate(transform(Functions.length()).check(gt(3))))))
+                CurriedFunctions.toCurriedContextPredicate(transform(Functions.length()).check(gt(3))))))
         .expectedExceptionClass(ComparisonFailure.class)
         /* Test Case Specific Check*/
         .addCheck(numberOfActualSummariesIsEqualTo(
@@ -64,7 +64,7 @@ public class StreamNoneMatchPredicateTest extends PropertyBasedTestBase {
         Stream.of("Hi", "hello", "world", null),
         transform(nest(asList("1", "2", "o")))
             .check(noneMatch(
-                Experimentals.toCurriedContextPredicate(transform(length()).check(gt(3))))))
+                CurriedFunctions.toCurriedContextPredicate(transform(length()).check(gt(3))))))
         //      |                         |         |                         |
         //      |                         |         |                         |
         //     (1)                       (2)       (3)                       (4),
@@ -86,7 +86,7 @@ public class StreamNoneMatchPredicateTest extends PropertyBasedTestBase {
     return new TestCase.Builder.ForThrownException<Stream<?>, ComparisonFailure>(
         Stream.of("Hi", "hello", "world"),
         transform(nest(asList("1", "2", "o")))
-            .check(noneMatch(Experimentals.toCurriedContextPredicate(stringEndsWith(), 0, 1))))
+            .check(noneMatch(CurriedFunctions.toCurriedContextPredicate(stringEndsWith(), 0, 1))))
         .expectedExceptionClass(ComparisonFailure.class)
         /* Test Case Specific Check*/
         .addCheck(numberOfActualSummariesIsEqualTo(
@@ -105,7 +105,7 @@ public class StreamNoneMatchPredicateTest extends PropertyBasedTestBase {
   public static TestCase<Stream<?>, ComparisonFailure> givenStreamPredicate$hello_b_e_4$_whenUnexpectedValue_thenComparisonFailure2() {
     return new TestCase.Builder.ForThrownException<Stream<?>, ComparisonFailure>(Stream.of(null, "Hi", "hello", "world", null),
         transform(nest(asList("1", "2", "o"))).check(noneMatch(
-            Experimentals.toCurriedContextPredicate(
+            CurriedFunctions.toCurriedContextPredicate(
                 transform(
                     Printables.function("throwIntentionalError", (Function<String, Integer>) s -> {
                       throw new IntentionalError();
