+++ src/test/java/com/github/dakusui/ut/thincrest/ut/PredicatesTest.java
@@ -1,5 +1,6 @@
 package com.github.dakusui.ut.thincrest.ut;
 
+import com.github.dakusui.pcond.experimentals.cursor.Cursors;
 import com.github.dakusui.pcond.forms.Functions;
 import com.github.dakusui.pcond.forms.Predicates;
 import com.github.dakusui.shared.utils.ut.TestBase;
@@ -339,7 +340,7 @@ public class PredicatesTest {
     public void givenSomeFoundSomeNot$whenFindString$thenFailed() {
       String text = "Gallia est omnis divisa in partes tres, quarum unum incolunt Belgae, aliam Acquitanii, tertiam nostra Galli Appellantur. De Bello Gallicco.";
       try {
-        TestAssertions.assertThat(text, Predicates.findSubstrings("Gallia", "quarum", "Belgium", "nostra", "De", "Gallia", "Gallicco"));
+        TestAssertions.assertThat(text, Cursors.findSubstrings("Gallia", "quarum", "Belgium", "nostra", "De", "Gallia", "Gallicco"));
       } catch (ComparisonFailure e) {
         e.printStackTrace();
         throw e;
@@ -350,7 +351,7 @@ public class PredicatesTest {
     public void givenAllFound$whenFindString$thenPassed() {
       String text = "Gallia est omnis divisa in partes tres, quarum unum incolunt Belgae, aliam Acquitanii, tertiam nostra Galli Appellantur. De Bello Gallicco.";
       try {
-        TestAssertions.assertThat(text, Predicates.findSubstrings("Gallia", "quarum", "Belgae", "nostra", "De", "Gallicco"));
+        TestAssertions.assertThat(text, Cursors.findSubstrings("Gallia", "quarum", "Belgae", "nostra", "De", "Gallicco"));
       } catch (ComparisonFailure e) {
         e.printStackTrace();
         throw e;
@@ -361,7 +362,7 @@ public class PredicatesTest {
     public void givenSomeFoundSomeNotFound$whenFindRegexes$thenFailed() {
       String text = "Gallia est omnis divisa in partes tres, quarum unum incolunt Belgae, aliam Acquitanii, tertiam nostra Galli Appellantur";
       try {
-        TestAssertions.assertThat(text, Predicates.findRegexes("Gall.a", "quar.m", "Belgium", "nostr(um|a)"));
+        TestAssertions.assertThat(text, Cursors.findRegexes("Gall.a", "quar.m", "Belgium", "nostr(um|a)"));
       } catch (ComparisonFailure e) {
         e.printStackTrace();
         throw e;
@@ -372,7 +373,7 @@ public class PredicatesTest {
     public void givenAllFound$whenFindRegexes$thenPassed() {
       String text = "Gallia est omnis divisa in partes tres, quarum unum incolunt Belgae, aliam Acquitanii, tertiam nostra Galli Appellantur";
       try {
-        TestAssertions.assertThat(text, Predicates.findRegexes("Gall.a", "quar.m", "Belg.+e,", "nostr(um|a)"));
+        TestAssertions.assertThat(text, Cursors.findRegexes("Gall.a", "quar.m", "Belg.+e,", "nostr(um|a)"));
       } catch (ComparisonFailure e) {
         e.printStackTrace();
         throw e;
@@ -386,7 +387,7 @@ public class PredicatesTest {
       List<String> list = asList("Hello", "world", "", "everyone", "quick", "brown", "fox", "runs", "forever");
       list.forEach(System.out::println);
       TestAssertions.assertThat(list,
-          Predicates.findElements(
+          Cursors.findElements(
               Predicates.isEqualTo("world"),
               Predicates.isEqualTo("cat"), Predicates.isEqualTo("organization"), Predicates.isNotNull(), Predicates.isEqualTo("fox"), Predicates.isEqualTo("world")));
     }
@@ -396,7 +397,7 @@ public class PredicatesTest {
       List<String> list = asList("Hello", "world", "", "everyone", "quick", "brown", "fox", "runs", "forever");
       list.forEach(System.out::println);
       TestAssertions.assertThat(list,
-          Predicates.findElements(
+          Cursors.findElements(
               Predicates.isEqualTo("world"),
               Predicates.isNotNull(),
               Predicates.isEqualTo("fox")));
